+++ release-ci.sh
@@ -25,7 +25,7 @@ fi
 
 echo "Attempting to release $LOCAL_VERSION (old version $BINTRAY_VERSION)"
 
-./gradlew -PreleaseVersion=$LOCAL_VERSION clean assemble \
+./gradlew -PreleaseVersion=$LOCAL_VERSION clean javadocJar assemble \
     :http4k-core:bintrayUpload \
     :http4k-aws:bintrayUpload \
     :http4k-cloudnative:bintrayUpload \