+++ http4k-cloudnative/src/main/kotlin/org/http4k/cloudnative/env/Environment.kt
@@ -1,24 +1,34 @@
 package org.http4k.cloudnative.env
 
 import org.http4k.core.Uri
-import org.http4k.lens.*
+import org.http4k.core.with
+import org.http4k.lens.BiDiLensSpec
+import org.http4k.lens.Lens
+import org.http4k.lens.LensGet
+import org.http4k.lens.LensSet
+import org.http4k.lens.ParamMeta
+import org.http4k.lens.duration
+import org.http4k.lens.int
 import java.io.File
 import java.io.Reader
-import java.util.*
+import java.time.Duration
+import java.util.Properties
 
 /**
- * This models the runtime environment of the shell where the app is running
+ * This models the runtime environment of the shell where the app is running. Optionally pass a separator to use for multi-values
+ * otherwise a standard comma is used - this means you MUST override the separator if you have single values which contain commas, otherwise
+ * singular environment keys will just retrieve the first value.
  */
-data class Environment private constructor(private val env: Map<String, String>) {
+data class Environment private constructor(private val contents: Map<String, String>, val separator: String = ",") {
     internal operator fun <T> get(key: Lens<Environment, T>) = key(this)
-    internal operator fun get(key: String): String? = env[key] ?: env[key.convertFromKey()]
-    internal operator fun set(key: String, value: String) = Environment(env + (key.convertFromKey() to value))
+    internal operator fun get(key: String): String? = contents[key] ?: contents[key.convertFromKey()]
+    internal operator fun set(key: String, value: String) = Environment(contents + (key.convertFromKey() to value))
 
     /**
      * Overlays the configuration set in this Environment on top of the values in the passed Environment.
      * Used to chain: eg. Local File -> System Properties -> Env Properties -> Defaults
      */
-    infix fun overrides(that: Environment) = Environment(that.env + env)
+    infix fun overrides(that: Environment) = Environment(that.contents + contents, separator)
 
     companion object {
         val EMPTY = from()
@@ -58,9 +68,18 @@ data class Environment private constructor(private val env: Map<String, String>)
     }
 }
 
+/**
+ * This models the key used to get a value out of the  Environment using the standard Lens mechanic. Note that if your values contain commas, either use a
+ * EnvironmentKey.(mapping).multi.required()/optional()/defaulted() to retrieve the entire list, or override the comma separator in your initial Environment.
+ */
 object EnvironmentKey : BiDiLensSpec<Environment, String>("env", ParamMeta.StringParam,
-    LensGet { name, target -> target[name]?.let { listOf(it) } ?: emptyList() },
-    LensSet { name, values, target -> values.fold(target) { acc, next -> acc.set(name, next) } }
+    LensGet { name, target -> target[name]?.split(target.separator) ?: emptyList() },
+    LensSet { name, values, target ->
+        values.fold(target) { acc, next ->
+            val existing = acc[name]?.let { listOf(it) } ?: emptyList()
+            acc.set(name, (existing + next).joinToString(target.separator))
+        }
+    }
 ) {
     object k8s {
         operator fun <T> invoke(fn: EnvironmentKey.k8s.() -> T): T = fn(this)
@@ -81,3 +100,14 @@ object EnvironmentKey : BiDiLensSpec<Environment, String>("env", ParamMeta.Strin
 }
 
 internal fun String.convertFromKey() = replace("_", "-").replace(".", "-").toLowerCase()
+
+fun main(args: Array<String>) {
+    val required = EnvironmentKey.duration().multi.required("FOO")
+    println(
+        required(
+            Environment.ENV.with(
+                required.of(listOf(Duration.ofHours(1), Duration.ofHours(2)))
+            )
+        )
+    )
+}